fix auth sock overwrite on reconnect

+++ .zshrc
@@ -27,7 +27,7 @@ alias lxs-stop='lxs-stop -n'
 stty -ixon
 
 # fix ssh forwarding for tmux
-if [[ -n "$SSH_TTY" && -S "$SSH_AUTH_SOCK" && ! -h "$SSH_AUTH_SOCK" ]]; then
+if [[ -n "$SSH_TTY" && -S "$SSH_AUTH_SOCK" && ! -h "$SSH_AUTH_SOCK" && ! -e ~/.ssh/ssh_auth_sock ]]; then
     ln -sf $SSH_AUTH_SOCK ~/.ssh/ssh_auth_sock
 fi
 export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
